+++ src/style.rs
@@ -42,6 +42,8 @@ impl From<Stroke> for Style {
 /// This is useful for methods that would like to accept draw styles by reference. Defining
 /// the type as `impl<Into<DrawRef>>` allows accepting types like `&Stroke` or `Fill`
 /// directly without cloning or allocating.
+#[allow(variant_size_differences)] // We don't expect this enum to be operated on in bulk.
+#[derive(Debug, Copy, Clone)]
 pub enum StyleRef<'a> {
     /// Filled draw operation.
     Fill(Fill),
@@ -49,7 +51,7 @@ pub enum StyleRef<'a> {
     Stroke(&'a Stroke),
 }
 
-impl<'a> StyleRef<'a> {
+impl StyleRef<'_> {
     /// Converts the reference to an owned draw.
     #[must_use]
     pub fn to_owned(&self) -> Style {