+++ CHANGELOG.md
@@ -17,12 +17,12 @@ This release has an [MSRV] of 1.70.
 
 ### Added
 
-- Breaking change: An `alpha` multiplier to `Image` ([#40][] by [@DJMcNab][])
+- Breaking: An `alpha` multiplier to `Image` ([#40][] by [@DJMcNab][])
 - `mint` feature to enable `mint` support in kurbo ([#46][] by [@waywardmonkeys][])
 
 ### Changed
 
-...
+- Breaking: Mark `Format` as `#[non_exhaustive]` ([#47][] by [@DJMcNab][])
 
 ### Fixed
 
@@ -47,6 +47,7 @@ This release has an [MSRV] of 1.70.
 [#26]: https://github.com/linebender/peniko/pull/26
 [#40]: https://github.com/linebender/peniko/pull/40
 [#46]: https://github.com/linebender/peniko/pull/46
+[#47]: https://github.com/linebender/peniko/pull/47
 
 [@DJMcNab]: https://github.com/DJMcNab
 [@ratmice]: https://github.com/ratmice